+++ src/traversable_types.mli
@@ -36,9 +36,10 @@ open Base
 
 (** {3 Inner-traversal signatures}
 
-    These signatures form the inner body of the [On_monad] functor in the
-    main signatures. They all have names ending with [_on_monad], and assume
-    the existence of a monad [M].
+    These signatures form the inner body of the [On] functors in the main
+    signatures. They all have names ending with [_on_applicative] or
+    [_on_monad], and assume the existence of a applicative functor or monad
+    [M].
 
     While they aren't interesting on their own, they do contain (in slightly
     abstract form) the specific functions needed to build, and provided on
@@ -46,34 +47,34 @@ open Base
 
 (** {4:omgeneric The generic signatures}
 
-    As with {{!Mappable} Mappable}, we define some signatures for traversable
-    structures in an arity-generic way, then specialise them for arity-0 and
-    arity-1 types. *)
+    As with {!Mappable}, we define some signatures for traversable structures
+    in an arity-generic way, then specialise them for arity-0 and arity-1
+    types. *)
 
-(** [Basic_generic_on_monad] describes monadic traversal on either an arity-0
+(** [Basic_generic_on_applicative] describes traversal on either an arity-0
     or arity-1 type.
 
     - For arity-0 types, ['a t] becomes [t] and ['a elt] becomes [elt].
     - For arity-1 types, ['a t] becomes ['a t] and ['a elt] becomes ['a]. *)
-module type Basic_generic_on_monad = sig
+module type Basic_generic_on_applicative = sig
   (** [Generic] refers to the container type as ['a t], and the element type
       as ['a elt]; substitute [t]/[elt] (arity-0) or ['a t]/['a] (arity-1)
       accordingly below. *)
   include Generic_types.Generic
 
-  (** [M] is the monad over which we're fold-mapping. *)
-  module M : Monad.S
+  (** [M] is the applicative functor over which we're fold-mapping. *)
+  module M : Applicative.S
 
   val map_m : 'a t -> f:('a elt -> 'b elt M.t) -> 'b t M.t
-  (** [map_m c ~f] maps [f] over every [t] in [c], threading through monadic
-      state.
+  (** [map_m c ~f] maps [f] over every [t] in [c], threading through an
+      applicative functor.
 
       Example:
 
       {[
-        (* Travesty_base_exts.List adds monadic traversals to a list;
+        (* Travesty_base_exts.List adds applicative traversals to a list;
            With_errors (in S1_container) implements them on the On_error
-           monad. *)
+           applicative functor. *)
 
         let f x =
           Or_error.(if 0 < x then error_string "negative!" else ok x)
@@ -82,59 +83,93 @@ module type Basic_generic_on_monad = sig
       ]} *)
 end
 
-(** [Generic_on_monad] extends [Generic] to contain various derived
-    operators; we use it to derive the signatures of the various [On_monad]
+(** [Generic_on_applicative] extends [Generic] to contain various derived
+    operators; we use it to derive the signatures of the various [On_*]
     modules.
 
     - For arity-0 types, ['a t] becomes [t] and ['a elt] becomes [elt].
     - For arity-1 types, ['a t] becomes ['a t] and ['a elt] becomes ['a]. *)
+module type Generic_on_applicative = sig
+  include Basic_generic_on_applicative
+
+  val iter_m : 'a t -> f:('a elt -> unit M.t) -> unit M.t
+  (** [iter_m x ~f] iterates the computation [f] over [x], returning the
+      final applicative effect. *)
+end
+
+(** [Generic_on_monad] extends [Generic_on_applicative] to contain various
+    derived operators that require monads; we use it to derive the signatures
+    of the various [On_monad] modules.
+
+    - For arity-0 types, ['a t] becomes [t] and ['a elt] becomes [elt].
+    - For arity-1 types, ['a t] becomes ['a t] and ['a elt] becomes ['a]. *)
 module type Generic_on_monad = sig
-  include Basic_generic_on_monad
+  (** The monad on which these operators are defined. *)
+  module M : Monad.S
+
+  (** All applicative operators are available through lowering the monad to
+      an applicative functor. *)
+  include Generic_on_applicative with module M := Monad_exts.App(M)
 
   val fold_map_m :
        'a t
     -> f:('acc -> 'a elt -> ('acc * 'b elt) M.t)
     -> init:'acc
     -> ('acc * 'b t) M.t
-  (** [fold_map_m c ~f ~init] folds [f] monadically over every [t] in [c],
+  (** [fold_map_m c ~f ~init] folds [f] applicatively over every [t] in [c],
       threading through an accumulator with initial value [init]. *)
 
   val fold_m :
     'a t -> init:'acc -> f:('acc -> 'a elt -> 'acc M.t) -> 'acc M.t
-  (** [fold_m x ~init ~f] folds the monadic computation [f] over [x],
+  (** [fold_m x ~init ~f] folds the applicative computation [f] over [x],
       starting with initial value [init], and returning the final value
-      inside the monadic effect. *)
+      inside the applicative effect. *)
 
   val iter_m : 'a t -> f:('a elt -> unit M.t) -> unit M.t
-  (** [iter_m x ~f] iterates the monadic computation [f] over [x], returning
-      the final monadic effect. *)
+  (** [iter_m x ~f] iterates the computation [f] over [x], returning the
+      final applicative effect. *)
 
   val mapi_m : f:(int -> 'a elt -> 'b elt M.t) -> 'a t -> 'b t M.t
-  (** [mapi_m ~f x] behaves as [mapM], but also supplies [f] with the index
+  (** [mapi_m ~f x] behaves as [map_m], but also supplies [f] with the index
       of the element. This index should match the actual position of the
       element in the container [x]. *)
 end
 
 (** {4:ombasic Basic signatures} *)
 
-(** [Basic0_on_monad] is the inner signature of a monadic traversal over
+(** [Basic0_on_applicative] is the inner signature of a traversal over
     arity-0 types. *)
-module type Basic0_on_monad = sig
+module type Basic0_on_applicative = sig
   include Generic_types.S0
 
-  include Basic_generic_on_monad with type 'a t := t and type 'a elt := elt
+  include
+    Basic_generic_on_applicative with type 'a t := t and type 'a elt := elt
 end
 
-(** [Basic1_on_monad] is the inner signature of a monadic traversal over
+(** [Basic1_on_applicative] is the inner signature of a traversal over
     arity-1 types. *)
-module type Basic1_on_monad = sig
+module type Basic1_on_applicative = sig
   (** The type of the container to map over. *)
   type 'a t
 
-  include Basic_generic_on_monad with type 'a t := 'a t and type 'a elt := 'a
+  include
+    Basic_generic_on_applicative with type 'a t := 'a t and type 'a elt := 'a
 end
 
 (** {4:oms Expanded signatures} *)
+
+(** [S1_on_applicative] extends [Generic_on_applicative] with functionality
+    that only works on arity-1 containers. *)
+module type S1_on_applicative = sig
+  type 'a t
+
+  include Generic_on_applicative with type 'a t := 'a t and type 'a elt := 'a
+
+  val sequence_m : 'a M.t t -> 'a t M.t
+  (** [sequence_m x] lifts a container of applicatives [x] to an applicative
+      containing a container, by sequencing the applicative effects from left
+      to right. *)
+end
 
 (** [S1_on_monad] extends [Generic_on_monad] with functionality that only
     works on arity-1 containers. *)
@@ -144,24 +179,27 @@ module type S1_on_monad = sig
   include Generic_on_monad with type 'a t := 'a t and type 'a elt := 'a
 
   val sequence_m : 'a M.t t -> 'a t M.t
-  (** [sequence_m x] lifts a container of monads [x] to a monad containing a
-      container, by sequencing the monadic effects from left to right. *)
+  (** [sequence_m x] lifts a container of applicatives [x] to an applicative
+      containing a container, by sequencing the applicative effects from left
+      to right. *)
 end
 
 (** {3 Basic signatures}
 
-    Any traversable type can be turned into a Core container, using the
-    monadic fold to implement all container functionality. The unified
-    signature of a container with monadic traversals is {{!S0} S0} (arity 0)
-    or {{!S1} S1} (arity 1).
+    Any traversable type can be turned into a Base container, using the
+    applicative fold to implement all container functionality. The unified
+    signature of a container with traversals is {!S0} (arity 0) or {!S1}
+    (arity 1).
 
-    To satisfy these signatures for new types, implement {{!Basic0} Basic0}
-    or {{!Basic1} Basic1}, and use the corresponding [MakeN] functor.
+    To satisfy these signatures for new types, implement {!Basic0} or
+    {!Basic1}, and use the corresponding [MakeN] functor. Note that you must
+    supply an [On] module in terms of applicatives, but not an [On_monad]
+    module; previous versions of Travesty required the opposite.
 
     For types that are _already_ Base containers, or types where custom
     implementation of the Base signature are desired, implement
-    {!Basic0_container} or {!Basic1_container}, and use the
-    [MakeN_container] functors. *)
+    {!Basic0_container} or {!Basic1_container}, and use the [MakeN_container]
+    functors. *)
 
 (** {4:basic For modules without a [Container] implementation} *)
 
@@ -174,9 +212,12 @@ module type Basic0 = sig
   (** [Elt] contains the element type, which must have equality. *)
   module Elt : Equal.S
 
-  (** [On_monad] implements monadic traversal for a given monad [M]. *)
-  module On_monad (M : Monad.S) :
-    Basic0_on_monad with type t := t and type elt := Elt.t and module M := M
+  (** [On] implements applicative traversal for a given applicative [M]. *)
+  module On (M : Applicative.S) :
+    Basic0_on_applicative
+      with type t := t
+       and type elt := Elt.t
+       and module M := M
 end
 
 (** [Basic1] is the minimal signature that traversable containers of arity 1
@@ -185,25 +226,23 @@ module type Basic1 = sig
   (** The container type. *)
   type 'a t
 
-  (** [On_monad] implements monadic traversal for a given monad. *)
-  module On_monad (M : Monad.S) :
-    Basic1_on_monad with type 'a t := 'a t and module M := M
+  (** [On] implements traversal for a given applicative functor. *)
+  module On (M : Applicative.S) :
+    Basic1_on_applicative with type 'a t := 'a t and module M := M
 end
 
 (** {4:bcon For modules with a [Container] implementation} *)
 
-(** [Basic0_container] combines {!Basic0} and the Base container
-    signature, and is used for extending existing containers into
-    {!S0_container}s. *)
+(** [Basic0_container] combines {!Basic0} and the Base container signature,
+    and is used for extending existing containers into {!S0_container}s. *)
 module type Basic0_container = sig
   include Basic0
 
   include Container.S0 with type t := t and type elt := Elt.t
 end
 
-(** [Basic1_container] combines {!Basic1} and the Base container
-    signature, and is used for extending existing containers into
-    {!S1_container} s. *)
+(** [Basic1_container] combines {!Basic1} and the Base container signature,
+    and is used for extending existing containers into {!S1_container} s. *)
 module type Basic1_container = sig
   include Basic1
 
@@ -217,8 +256,16 @@ end
 module type Generic = sig
   include Generic_types.Generic
 
-  (** [On_monad] implements monadic traversal operators for a given monad
-      [M]. *)
+  (** [On] implements traversal operators for a given applicative [M]. *)
+  module On (M : Applicative.S) :
+    Generic_on_applicative
+      with type 'a t := 'a t
+       and type 'a elt := 'a elt
+       and module M := M
+
+  (** [On_monad] implements traversal operators for a given monad [M].
+      Compared to [On(Monad_exts.App(M))], this adds various derived
+      operators available only for monads. *)
   module On_monad (M : Monad.S) :
     Generic_on_monad
       with type 'a t := 'a t
@@ -228,7 +275,7 @@ module type Generic = sig
   (** We can do generic container operations. *)
   include Container.Generic with type 'a t := 'a t and type 'a elt := 'a elt
 
-  (** We can do non-monadic mapping operations. *)
+  (** We can do non-applicative mapping operations. *)
   include
     Mappable_types.Generic with type 'a t := 'a t and type 'a elt := 'a elt
 
@@ -241,7 +288,7 @@ module type Generic = sig
   (** [mapi ~f t] maps [f] across [t], passing in an increasing position
       counter. *)
 
-  (** [With_errors] specialises [On_monad] to the error monad. *)
+  (** [With_errors] specialises [On_applicative] to the error applicative. *)
   module With_errors :
     Generic_on_monad
       with type 'a t := 'a t
@@ -266,14 +313,19 @@ module type S0 = sig
 end
 
 (** [S1] is a generic interface for arity-1 traversable containers. It also
-    includes the extensions from {{!Mappable} Mappable}. *)
+    includes the extensions from {!Mappable}. *)
 module type S1 = sig
   (** ['a t] is the type of the container, parametrised over the element type
       ['a]. *)
   type 'a t
 
-  (** [On_monad] implements monadic folding and mapping operators for a given
-      monad [M], including arity-1 specific operators. *)
+  (** [On] implements applicative folding and mapping operators for a given
+      applicative [M], including arity-1 specific operators. *)
+  module On (M : Applicative.S) :
+    S1_on_applicative with type 'a t := 'a t and module M := M
+
+  (** [On_monad] implements applicative folding and mapping operators for a
+      given monad [M], including arity-1 specific operators. *)
   module On_monad (M : Monad.S) :
     S1_on_monad with type 'a t := 'a t and module M := M
 
@@ -285,6 +337,7 @@ module type S1 = sig
     Generic
       with type 'a t := 'a t
        and type 'a elt := 'a
+       and module On := On
        and module On_monad := On_monad
        and module With_errors := With_errors